refactoring looks good so far
