+++ entropy/file_utils.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+
+def file_exits(path):
+    return Path(path).exists()
